# openapi.yaml
openapi: 3.0.3
info:
  title: Joinly API
  version: 0.1.0
  description: API REST del MVP de eventos locales. Los recursos de producto requieren un JWT Bearer de Supabase Auth.
servers:
  - url: http://localhost:8080/api/v1
    description: Desarrollo local
security:
  - bearerAuth: []
paths:
...
  /me/events:
    get:
      summary: Lista eventos propios
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}, {name: status, in: query, schema: {type: string, enum: [published, cancelled, closed]}}]
      responses:
        '200': {description: Eventos, content: {application/json: {schema: {$ref: '#/components/schemas/EventPage'}}}}
  /events:
    post:
      summary: Publica un evento
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/EventInput'}}}
      responses:
        '201': {description: Evento publicado, headers: {ETag: {$ref: '#/components/headers/ETag'}}, content: {application/json: {schema: {$ref: '#/components/schemas/EventDetail'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /events/search:
    post:
      summary: Descubre eventos futuros por radio
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/EventSearch'}}}
      responses:
        '200': {description: Eventos visibles sin ubicación exacta, content: {application/json: {schema: {$ref: '#/components/schemas/EventSearchPage'}}}}
        '400': {$ref: '#/components/responses/ValidationError'}
  /events/{eventId}:
    parameters: [{$ref: '#/components/parameters/EventId'}]
    get:
      summary: Obtiene un evento visible
      responses:
        '200': {description: Evento, headers: {ETag: {$ref: '#/components/headers/ETag'}}, content: {application/json: {schema: {$ref: '#/components/schemas/EventDetail'}}}}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      summary: Edita un evento propio
      parameters: [{$ref: '#/components/parameters/IfMatch'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/EventPatch'}}}
      responses:
        '200': {description: Evento actualizado, headers: {ETag: {$ref: '#/components/headers/ETag'}}, content: {application/json: {schema: {$ref: '#/components/schemas/EventDetail'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '428': {$ref: '#/components/responses/PreconditionRequired'}
  /events/{eventId}/cancellation:
    parameters: [{$ref: '#/components/parameters/EventId'}]
    post:
      summary: Cancela un evento propio antes de su inicio
      requestBody:
        content: {application/json: {schema: {type: object, properties: {reason: {type: string, maxLength: 4000}}}}}
      responses:
        '204': {description: Evento cancelado}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /blocks:
    get:
      summary: Lista bloqueos creados por la persona autenticada
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses: {'200': {description: Bloqueos, content: {application/json: {schema: {$ref: '#/components/schemas/CursorPage'}}}}}
    post:
      summary: Crea un bloqueo recíproco lógico
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [blockedUserId], properties: {blockedUserId: {$ref: '#/components/schemas/Uuid'}}}}}
      responses:
        '201': {description: Bloqueo creado}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /blocks/{blockedUserId}:
    parameters: [{$ref: '#/components/parameters/BlockedUserId'}]
    delete:
      summary: Elimina un bloqueo propio
      responses: {'204': {description: Bloqueo eliminado}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  headers:
    ETag: {schema: {type: string}, description: Versión opaca del recurso}
  parameters:
    EventId: {name: eventId, in: path, required: true, schema: {$ref: '#/components/schemas/Uuid'}}
...
    BlockedUserId: {name: blockedUserId, in: path, required: true, schema: {$ref: '#/components/schemas/Uuid'}}
    Cursor: {name: cursor, in: query, schema: {type: string}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
    IfMatch: {name: If-Match, in: header, required: false, schema: {type: string}}
...
  responses:
...
    Forbidden: {description: Operación no permitida, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    NotFound: {description: Recurso inexistente u oculto, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    Conflict: {description: Regla de negocio incumplida, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    ValidationError: {description: Solicitud inválida, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    PreconditionFailed: {description: ETag desactualizado, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
    PreconditionRequired: {description: Falta If-Match, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
  schemas:
    Uuid: {type: string, format: uuid}
    GeoPoint:
      type: object
      required: [type, coordinates]
      properties: {type: {type: string, enum: [Point]}, coordinates: {type: array, minItems: 2, maxItems: 2, items: {type: number}}}
    PublicProfile: {type: object, required: [id, alias], properties: {id: {$ref: '#/components/schemas/Uuid'}, alias: {type: string}}}
...
    EventInput: {type: object, required: [title, description, category, startsAt, durationMinutes, exactLocation, accessMode], properties: {title: {type: string, minLength: 3, maxLength: 120}, description: {type: string, maxLength: 4000}, notes: {type: string, maxLength: 4000}, category: {type: string, enum: [sportWellbeing, cultureLeisure, learning, communityVolunteering, pets, networking]}, startsAt: {type: string, format: date-time}, durationMinutes: {type: integer, minimum: 15, maximum: 1440}, exactLocation: {$ref: '#/components/schemas/GeoPoint'}, accessMode: {type: string, enum: [direct, approval, privateInvitation]}, capacity: {type: integer, minimum: 1, nullable: true}}}
    EventPatch: {type: object, minProperties: 1, properties: {title: {type: string, minLength: 3, maxLength: 120}, description: {type: string, maxLength: 4000}, notes: {type: string, maxLength: 4000}, category: {type: string, enum: [sportWellbeing, cultureLeisure, learning, communityVolunteering, pets, networking]}, startsAt: {type: string, format: date-time}, durationMinutes: {type: integer, minimum: 15, maximum: 1440}, exactLocation: {$ref: '#/components/schemas/GeoPoint'}, accessMode: {type: string, enum: [direct, approval, privateInvitation]}, capacity: {type: integer, minimum: 1, nullable: true}}}
    EventSearch: {type: object, required: [origin, radiusMeters], properties: {origin: {$ref: '#/components/schemas/GeoPoint'}, radiusMeters: {type: integer, minimum: 100, maximum: 50000}, categories: {type: array, items: {type: string}}, cursor: {type: string, nullable: true}, limit: {type: integer, minimum: 1, maximum: 50, default: 20}}}
    EventDiscovery: {type: object, required: [id, title, description, category, startsAt, durationMinutes, accessMode, confirmedCount, availability, approximateArea, creator], properties: {id: {$ref: '#/components/schemas/Uuid'}, title: {type: string}, description: {type: string}, category: {type: string}, startsAt: {type: string, format: date-time}, durationMinutes: {type: integer}, accessMode: {type: string}, capacity: {type: integer, nullable: true}, confirmedCount: {type: integer}, availability: {type: string, enum: [available, full]}, approximateArea: {type: string}, distanceMeters: {type: integer}, creator: {$ref: '#/components/schemas/PublicProfile'}}}
    EventDetail: {allOf: [{$ref: '#/components/schemas/EventDiscovery'}, {type: object, properties: {notes: {type: string}, updatedAt: {type: string, format: date-time}, myParticipation: {type: string, enum: [pending, confirmed, rejected, abandoned]}, exactLocation: {$ref: '#/components/schemas/GeoPoint'}, confirmedParticipants: {type: array, items: {$ref: '#/components/schemas/PublicProfile'}}}}]}
...
    CursorPage: {type: object, required: [items, page], properties: {items: {type: array, items: {type: object}}, page: {type: object, required: [nextCursor], properties: {nextCursor: {type: string, nullable: true}}}}}
    EventPage: {allOf: [{$ref: '#/components/schemas/CursorPage'}]}
    EventSearchPage: {allOf: [{$ref: '#/components/schemas/CursorPage'}, {type: object, properties: {suggestedRadiusMeters: {type: integer}}}]}
...
    Problem: {type: object, required: [type, title, status, code], properties: {type: {type: string, format: uri}, title: {type: string}, status: {type: integer}, code: {type: string}, detail: {type: string}, fields: {type: object, additionalProperties: {type: string}}}}